Move `GASDSignature984` to feature types header
