+++ common/src/feature_types.cpp
@@ -164,4 +164,12 @@ namespace pcl
       os << (i == 0 ? "(" : "") << p.histogram[i] << (i < 511 ? ", " : ")");
     return (os);
   }
+
+  std::ostream&
+  operator << (std::ostream& os, const GASDSignature984& p)
+  {
+    for (int i = 0; i < 984; ++i)
+      os << (i == 0 ? "(" : "") << p.histogram[i] << (i < 983 ? ", " : ")");
+    return (os);
+  }
 } // namespace pcl
