+++ common/src/point_types.cpp
@@ -242,14 +242,6 @@ namespace pcl
   }
 
   std::ostream&
-  operator << (std::ostream& os, const GASDSignature984& p)
-  {
-    for (int i = 0; i < 984; ++i)
-      os << (i == 0 ? "(" : "") << p.histogram[i] << (i < 983 ? ", " : ")");
-    return (os);
-  }
-
-  std::ostream&
   operator << (std::ostream& os, const GASDSignature7992& p)
   {
     for (int i = 0; i < 7992; ++i)
